backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
mysqlclient==2.2.0
alembic==1.13.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aioredis==2.0.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
uuid==1.30
websockets==12.0
asyncpg==0.29.0
httpx==0.25.2
structlog==23.2.0
prometheus-client==0.19.0
python-json-logger==2.0.7